requirements.txt
streamlit
openai
azure-ai-textanalytics
pandas
requests
python-dotenv
plotly